Refactored text functions out of util module
This move was needed for avoiding circular imports from upcoming
refactoring of logging.
